[OPT]update to zy_compass_4.1: 1. support subgraph for MoE structure 2. update svd_quant for fpx quantization 3. add fpx quantization support 4. add operator: Shape,ConstantOfShape, RoPE 5. refine codes and fix bugs

+++ AIPUBuilder/Optimizer/ops/upsamplebyindex.py
@@ -7,7 +7,7 @@ from AIPUBuilder.Optimizer.framework import *
 
 @op_register(OpType.UpsampleByIndex)
 def upsamplebyindex(self, *args):
-    values = self.inputs[0].betensor
+    values = self.inputs[0].betensor.float()
     argmax = self.inputs[1].betensor.long()
 
     current_batch = values.shape[0]
